Refactor comment PATCH tests
- refactor comments flag tests
- refactor replies flag tests
- refactor replies vote tests
- refactor replies edit tests
- add file to test GET comments votes by query parameters
- drop old comment patch tests file

See #241
